sink writes cloudevents to the database
Need to make this work with the full db schema, clean up the code, and
figure out a better solution for connecting to the database. This works
though. You just need to deploy the database before installing
kubearchive and update the host to use the IP of the database service

Signed-off-by: skoved <skoved@redhat.com>