Sink writes cluster name to datbase and updates entries
The sink reads the name of the cluster from an environment variable and
writes that value for entries that it addes to the database. The sink
will also check if a resource has already been written to the database
and update the entry if the resource has already been written. Otherwise
it will create a new entry

Signed-off-by: skoved <skoved@redhat.com>