sink does not try to delete k8s objects twice
The after the sink deletes a kubernetes object, it will update the entry
in the database for that resource with timestamp of when the resource was
deleted. Bumped the version of knative-eventing to v1.15.0 so that the
operator can apply a filter to all ApiServerSources so that they do not
send delete type cloudevents.

Signed-off-by: skoved <skoved@redhat.com>