Sink archives and deletes resources based on CEL expressions from ConfigMap (#286)
Sink archives and deletes resources on the cluster depending on if they
cause a cel expression to return to true. The cel expressions that are run
depend on the namespace and apiVersion of the k8s object. When the ConfigMap
is changed, the sink will respond to those changes by creating, deleting and
updating the filters it has stored with the new cel expressions from the
ConfigMap.

Signed-off-by: skoved <skoved@redhat.com>