sink can archive objects when they are deleted
The sink now checks if the received cloudevent's type is 'delete'. If it is, the sink checks if the object contained in
the cloudevent meets the criteria for archiveOnDelete. If the object meets that criteria, the object is written to the
database. If the object does not meet the criteria, the sink ignores the message. If the cloudevent's type is not
'delete', the sink handles the cloudevent normally. The sink also will not a attempt to compile a CEL expression string
if it is empty to avoid CEL compilation errors.

Signed-off-by: skoved <skoved@redhat.com>