kubearchive uses knative brokers to handle event delivery
KubeArchive uses a knative eventing broker to receive cloudevents from
ApiServerSource and uses a knative eventing trigger to subscribe the sink to
all events from the broker. The broker is configured to retry the delivery of
cloudevents if the initial delivery fails. If all retries fail, the broker sends
the cloudevent an other broker so that the cloudevents are not lost.

Signed-off-by: skoved <skoved@redhat.com>