run integrations tests in parallel
TestEmptyKAC, TestNonEmptyKAC, and TestDatabaseConnection cannot be run in parallel. TestEmptyKAC and TestNonEmptyKAC
cannot be run in parallel because they are affected by KACs created by other tests. TestDatabaseConnection cannot be run
in parallel because messing with the api server and sink database connections affect other tests. The code to port
forward the api server and run the log generators has been modified to allow each to be called concurrently. Functions
that write to stdout or stderr have been updated to call t.Log or t.Logf to make it easier to understand which test is
printing what output. Helper functions that write to stdout or stderr now call t.Helper to register themselves as helper
functions to make tracing errors easier.

Signed-off-by: skoved <skoved@redhat.com>