integration test action should fail when go test cmd fails (#918)
I forgot to add `set -o pipefail` in test/integration/run.sh. This cases the integration test action to report as passed
instead of failed when the `go test` command fails. Also unset errexit before running integration tests so that the test
logs can be processed and an action summary can be generated even on failure. Return code for the integration test's go
test command is saved to be used as the return code for the script.

Signed-off-by: skoved <skoved@redhat.com>