Old versions of resources do not get written to db
Add cluster_updated_ts column to the database which stores time stamp for the last update of the resource on the
cluster. Added a function to the sink to calculate the time that the resource was last updated using
metadata.managedFields. This time is then passed to the database.WriteResource. ResourceInserter methods now add a where
clause to the sql insert to prevent the on conflict clause from activating when cluster_updated_ts in the database is
more recent than the one provided in the insert.

Signed-off-by: skoved <skoved@redhat.com>