Refactor database to separate write and read functions (#921)
* Rename DBInterface and Database

To be consistent with the rest of the code.
Interfaces won't have Interface in suffix.
Default implementations will have Impl
in the suffix.

Signed-off-by: Marta Anon <manon@redhat.com>

* Segregate Database interfaces

DBWriter encapsulates the needed functions
for the sink
DBReader encapsulates the needed functions
for the api

Signed-off-by: Marta Anon <manon@redhat.com>

* Extract reader and writer logic

DBReader and DBWriter logic moved
to its own files to improve
readability.

Signed-off-by: Marta Anon <manon@redhat.com>

* Remove README.md as it has duplicated info

Signed-off-by: Marta Anon <manon@redhat.com>

---------

Signed-off-by: Marta Anon <manon@redhat.com>